requirements.txt
fastapi==0.135.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

sqlalchemy==2.0.36
psycopg2==2.9.9
alembic==1.13.1
pgvector==0.2.5

pydantic==2.12.5
pydantic-settings==2.1.0
python-dotenv==1.0.0

langchain-core==0.2.43
langsmith==0.1.112
langchain-openai==0.1.25
langgraph==0.2.14
openai==1.109.1

pika==1.4.0 

requests==2.32.5
httpx==0.28.1

beautifulsoup4==4.12.3
lxml==6.0.2
tenacity==8.2.3

PyPDF2==3.0.1
python-docx==1.1.0

jieba==0.42.1

nacos-sdk-python==0.1.11
neo4j==5.16.0
